Create iamge macro
To make the actions in the computer vision-based macro sequenced, you can store the state of each action in a variable and check the state before performing the action. For example, you can use a dictionary to map each target image to a corresponding action state, and set the state to True when the action is performed and False when the action is not performed. Here is an example of how you might modify the code to do this: